LayoutTests/webanimations/animation-effect.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check that the AnimationEffect interface is defined and does not expose a constructor.");

debug("The AnimationEffect interface is defined.");
shouldBeDefined("AnimationEffect");

debug("");
debug("The AnimationEffect interface cannot be constructed.");
shouldThrow("new AnimationEffect");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
