LayoutTests/webanimations/animation-effect-expected.txt

Check that the AnimationEffect interface is defined and does not expose a constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The AnimationEffect interface is defined.
PASS AnimationEffect is defined.

The AnimationEffect interface cannot be constructed.
PASS new AnimationEffect threw exception TypeError: function is not a constructor (evaluating 'new AnimationEffect').

PASS successfullyParsed is true

TEST COMPLETE
